+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8.26 2022-03-13
+
+* replace system_info package by native uname() call to detect the cpu architecture.
+* add uname() support- see [example](https://github.com/pezi/dart_periphery/blob/main/example/uname.dart). 
+
 ## 0.8.25-RC 2021-04-26
 
 * Add isolate support for SPI, I2C and Serial.